requirements.txt
PyGithub==2.5.0
requests==2.32.3

